chore(app) refacto adjustments
