+++ APP/lib/l10n/app_fr.arb
@@ -68,12 +68,14 @@
   "searchById": "Rechercher par ID",
   "filters": "Filtres",
   "clearAllFilters": "Effacer tout",
-  "noNodeFound": "Pas de node trouvé avec l'ID :",
+  "noNodeFound": "Pas de noeud trouvé avec l'ID :",
+  "nodeFound": "Noeud trouvé :",
   "expandAll" : "Développer tout",
   "collapseAll": "Réduire tout",
   "selectAll": "Sélectionner tout",
   "deselectAll": "Désélectionner tout",
   "nodePath": "Chemin du noeud :",
+  "toggleSelection": "Inverser sélection du noeud et de tous ses enfants",
 
   "@result": { "description": "Result Page" },
   "result": "Résultat",
