+++ APP/lib/pages/tenant_page.dart
@@ -3,18 +3,13 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:ogree_app/common/api_backend.dart';
 import 'package:ogree_app/common/appbar.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:ogree_app/common/definitions.dart';
-import 'package:ogree_app/common/popup_dialog.dart';
-import 'package:ogree_app/common/theme.dart';
 import 'package:ogree_app/models/tenant.dart';
 import 'package:ogree_app/pages/projects_page.dart';
 import 'package:ogree_app/widgets/tenants/api_stats_view.dart';
+import 'package:ogree_app/widgets/tenants/domain_view.dart';
 import 'package:ogree_app/widgets/tenants/locked_view.dart';
-import 'package:ogree_app/widgets/tenants/popups/domain_popup.dart';
 import 'package:ogree_app/widgets/tenants/docker_view.dart';
 import 'package:ogree_app/widgets/select_objects/treeapp_controller.dart';
-import 'package:ogree_app/widgets/select_objects/settings_view/settings_view.dart';
-import 'package:ogree_app/widgets/select_objects/tree_view/custom_tree_view.dart';
 import 'package:ogree_app/widgets/tenants/user_view.dart';
 
 class TenantPage extends StatefulWidget {
@@ -33,8 +28,6 @@ class _TenantPageState extends State<TenantPage> with TickerProviderStateMixin {
   late TabController _tabController;
   late final TreeAppController appController = TreeAppController();
   bool _isLocked = true;
-  bool _reloadDomains = false;
-  bool _isSmallDisplay = false;
   String _domainSearch = "";
 
   @override
@@ -47,7 +40,6 @@ class _TenantPageState extends State<TenantPage> with TickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
     final localeMsg = AppLocalizations.of(context)!;
-    _isSmallDisplay = IsSmallDisplay(MediaQuery.of(context).size.width);
     return Scaffold(
         backgroundColor: const Color.fromARGB(255, 238, 238, 241),
         appBar: myAppBar(context, widget.userEmail,
@@ -159,7 +151,7 @@ class _TenantPageState extends State<TenantPage> with TickerProviderStateMixin {
           : ApiStatsView(),
       _isLocked && widget.tenant != null
           ? LockedView(tenant: widget.tenant!, parentCallback: unlockView)
-          : domainView(localeMsg),
+          : DomainView(),
       _isLocked && widget.tenant != null
           ? LockedView(tenant: widget.tenant!, parentCallback: unlockView)
           : (_domainSearch.isEmpty
@@ -188,76 +180,6 @@ class _TenantPageState extends State<TenantPage> with TickerProviderStateMixin {
     setState(() {
       _isLocked = false;
     });
-  }
-
-  domainView(AppLocalizations localeMsg) {
-    return Stack(children: [
-      AppControllerScope(
-        controller: appController,
-        child: FutureBuilder<void>(
-          future: appController.init({},
-              namespace: Namespace.Organisational,
-              reload: _reloadDomains,
-              isTenantMode: true),
-          builder: (_, __) {
-            if (_reloadDomains) {
-              _reloadDomains = false;
-            }
-            if (appController.isInitialized) {
-              if (appController.treeController.roots.isEmpty) {
-                return Column(
-                  children: [
-                    Icon(
-                      Icons.warning_rounded,
-                      size: 50,
-                      color: Colors.grey.shade600,
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(top: 16),
-                      child: Text(
-                          "${AppLocalizations.of(context)!.noObjectsFound} :("),
-                    ),
-                  ],
-                );
-              }
-              return Stack(children: [
-                const CustomTreeView(isTenantMode: true),
-                _isSmallDisplay
-                    ? Container()
-                    : const Align(
-                        alignment: Alignment.topRight,
-                        child: Padding(
-                          padding: EdgeInsets.only(right: 16),
-                          child: SizedBox(
-                              width: 320,
-                              height: 116,
-                              child: Card(
-                                  child: SettingsView(isTenantMode: true))),
-                        ),
-                      ),
-              ]);
-            }
-            return const Center(child: CircularProgressIndicator());
-          },
-        ),
-      ),
-      Align(
-        alignment: Alignment.bottomRight,
-        child: Padding(
-          padding: const EdgeInsets.only(bottom: 20, right: 20),
-          child: ElevatedButton.icon(
-            onPressed: () =>
-                showCustomPopup(context, DomainPopup(parentCallback: () {
-              setState(() {
-                _reloadDomains = true;
-              });
-            })),
-            icon: const Icon(Icons.add),
-            label: Text("${localeMsg.create} ${localeMsg.domain}"),
-          ),
-        ),
-      ),
-    ]);
   }
 
   changeToUserView(String domain) {
