+++ APP/lib/widgets/select_objects/settings_view/settings_view.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:ogree_app/common/snackbar.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:ogree_app/common/theme.dart';
 
 import '../treeapp_controller.dart';
 import '../tree_view/tree_node.dart';
