+++ APP/lib/widgets/select_objects/settings_view/tree_filter.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:ogree_app/common/theme.dart';
 import 'package:ogree_app/widgets/select_objects/treeapp_controller.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
@@ -191,11 +192,8 @@ class _AutocompleteFilterState extends State<AutocompleteFilter> {
             return TextFormField(
               controller: textEditingController,
               focusNode: focusNode,
-              decoration: InputDecoration(
-                  enabled: widget.enabled,
-                  isDense: true,
-                  labelText: widget.param,
-                  labelStyle: const TextStyle(fontSize: 14)),
+              decoration: GetFormInputDecoration(true, widget.param,
+                  isEnabled: widget.enabled),
               onFieldSubmitted: (String value) {
                 if (widget.options.contains(value)) {
                   setState(() {
@@ -245,12 +243,12 @@ class _AutocompleteFilterState extends State<AutocompleteFilter> {
             );
           },
         ),
-        const SizedBox(height: 4),
         Wrap(
           spacing: 10,
           runSpacing: 10,
           children: getChips(_selectedOptions, context),
         ),
+        const SizedBox(height: 4),
       ],
     );
   }
@@ -258,7 +256,7 @@ class _AutocompleteFilterState extends State<AutocompleteFilter> {
   // One chip per selected filter
   List<Widget> getChips(List<String> nodes, BuildContext context) {
     List<Widget> chips = [];
-    nodes.forEach((value) {
+    for (var value in nodes) {
       chips.add(RawChip(
         onPressed: () {
           TreeAppController.of(context).filterTree(value, widget.paramLevel);
@@ -284,7 +282,7 @@ class _AutocompleteFilterState extends State<AutocompleteFilter> {
           color: ColorChip[widget.param],
         ),
       ));
-    });
+    }
     return chips;
   }
 }
